Added Observation type field and removed Carma23.
Changed the Carma 23 Checkbox to a dropdown box that lets you choose
between SINGLEPOL (default), DUALPOL, FULLPOL, and CARMA23. Added in
error checking between this field, the frequency field, and the hours
fields in order to make sure that people are entering sane values.

There is still a possibility of entering non-sane values, since I did
not do an exhaustive search of input values, however, I *think* it is
all correct.

Adjusted the 2012a latex template to match changes. The C23 box needed
to be slightly widened, and made the # Fields box slightly smaller.