CPSS MERGE - Rewrote the url dispatcher. Fixed a couple db bugs.
The URL dispatcher was convoluted, yet worked. Expanded to a rules
based system that defines clear tasks for each function. This is the
beginning of modifying the old proposal logic code to be small,
modular, and easily debugable. The whole proposal system will
require a large debugging session once this is done.

Its this logic code that needs to be cleaned up and updated to support
multiple active proposal templates at one time. For example, the
add proposal command needs to do a search on the current open
cycles and pick the correct template/tables to modify from that.

Also fixed a bug in the db about loading PDF files from disk.
Tidied up the stats page a bit. The rest of the stats cleaning
up will be in the db itself (the cycles.enddata specifically as
that determines the order the cycles appear on the statspage.