CPSS MERGE - finished the add proposal command.
This change is going to require further edits down the road for other
parts of the Template routine as it assumed the cycleinfo was not
dynamic, but rather set in the database options.

Changed many of the database routines to not query the tablenames, but
get passed them. The calling fuctions must be changed to have the cycleinfo
and pick the correct table from there. These will be changed as we run into
them.

These changes will need extensive debugging to ensure that there are no
unintended side effects.