CPSS MERGE - added in typechange and viewable proposals.
Proposals can now be viewed after they've been submitted.

Massively cleaned up some html commands in the template.

Added in the typechange command to make sure the html was being generated
correctly for uploading the justifications.

THIS WAS AN UGLY MESS.
Its still ugly and not as clean as it could be, but it works better than it
did previously.